[fix] applicationコントローラー内に認証のためのメソッドを移動
